// src/types/api.ts
// Generated by https://quicktype.io

import type { ReactNode } from "react";

export interface Product {
    images: any;
    sku: ReactNode;
    inventory: any;
    name: ReactNode;
    data: ProductData;
}

export interface ProductData {
    id:                     number;
    brand_id:               number;
    country_id:             number;
    product_id:             number;
    sku:                    string;
    name:                   string;
    list_price:             string;
    standard_dealer_price:  string;
    supplier_product_id:    string;
    length:                 number;
    width:                  number;
    height:                 number;
    weight:                 number;
    upc:                    null;
    superseded_sku:         null;
    status_id:              string;
    status:                 string;
    unit_of_measurement_id: number;
    has_map_policy:         boolean;
    sort:                   number;
    created_at:             string;
    updated_at:             string;
    published_at:           string;
    product_type:           string;
    mapp_price:             string;
    carb:                   null;
    propd1:                 null;
    propd2:                 null;
    prop_65_code:           string;
    prop_65_detail:         string;
    drop_ship_fee:          string;
    drop_ship_eligible:     boolean;
    images:                 Images;
    inventory:              Inventory;
}

export interface Images {
    data: Datum[];
}

export interface Datum {
    id:         number;
    domain:     string;
    path:       string;
    filename:   string;
    alt:        null;
    mime:       string;
    width:      number;
    height:     number;
    size:       number;
    signature:  string;
    created_at: string;
    updated_at: string;
}

export interface Inventory {
    data: InventoryData;
}

export interface InventoryData {
    id:            number;
    item_id:       number;
    sku:           string;
    ca_warehouse:  number;
    ga_warehouse:  number;
    id_warehouse:  number;
    in_warehouse:  number;
    pa_warehouse:  number;
    pa2_warehouse: number;
    tx_warehouse:  number;
    total:         number;
    created_at:    string;
    updated_at:    string;
}